src/main/webapp/app/entities/user-academic-info/delete/user-academic-info-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { UserAcademicInfoService } from '../service/user-academic-info.service';

import { UserAcademicInfoDeleteDialogComponent } from './user-academic-info-delete-dialog.component';

describe('UserAcademicInfo Management Delete Component', () => {
  let comp: UserAcademicInfoDeleteDialogComponent;
  let fixture: ComponentFixture<UserAcademicInfoDeleteDialogComponent>;
  let service: UserAcademicInfoService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [UserAcademicInfoDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(UserAcademicInfoDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(UserAcademicInfoDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(UserAcademicInfoService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
